requirements.txt
Django
Pillow
django-allauth
django-crispy-forms
django-jsonfield
django-modeltranslation
django-rosetta
django-summernote
django-constance
django-picklefield
django-dynamic-fixture
sorl-thumbnail
